complete responsive homepage, refactor of PLP using grid
