+++ src/components/home/Carousel/CarouselItem.tsx
