+++ src/components/product/ProductList.tsx
@@ -1,24 +1,29 @@
 import ProductItem from '@components/product/ProductItem';
 import { Models } from 'utilities-techsweave';
-import React, { useState } from 'react';
-import { Flex, Box, VStack } from '@chakra-ui/react';
+import React from 'react';
+import { Grid, GridItem, Box } from '@chakra-ui/react';
 
 const ProductList = (prop: { productList: Models.Tables.IProduct[] }) => {
   const { productList } = prop;
 
   return (
-    <VStack>
-      <Flex direction={['column', 'column', 'row', 'row']}>
-        <Box>
-          {productList.map((products) => (
+    <Box>
+      <Grid ml={['0', '0', '50', '50']} templateColumns={['repeat(1, 1fr)', 'repeat(1, 1fr)', 'repeat(4, 1fr)', 'repeat(4, 1fr)']} gap={10}>
+        {productList.map((products) => (
+          <GridItem
+            id={products.id}
+            tag="li"
+            style={{ listStyle: 'none' }}
+          >
             <ProductItem
               product={products}
               key={products.id}
             />
-          ))}
-        </Box>
-      </Flex>
-    </VStack>
+          </GridItem>
+        ))}
+      </Grid>
+    </Box>
+
   );
 };
