+++ src/pages/index.tsx
@@ -1,6 +1,6 @@
 import Layout from '@components/Layout';
 import React from 'react';
-import Carousel from '@components/home/Carousel';
+import Carousel from '@components/home/Carousel/Carousel';
 import RecentProduct from '@components/home/RecentProduct';
 import productMock from '@test/ProductMock';
 import { Box } from '@chakra-ui/react';
@@ -9,7 +9,7 @@ const indexPage = () => (
   <Layout title="EmporioLambda">
     <Box w='95%'>
       <Carousel product={productMock} />
-      <RecentProduct />
+      <RecentProduct product={productMock} />
     </Box>
   </Layout>
 );
